refactor: reduce the join-order rule's header to the rule itself
The header declared ten things, of which one -- the rule function -- had a
production consumer: OptimizerRulesFeature registers it and uses nothing
else. The other nine were the algorithm's internal steps, exposed only so
the tests could reach them, and the header's own includes handed every
caller the whole join-graph API along the way.

Move those steps into a unit owned by the rule's own directory, which the
tests include directly, and leave the header declaring the entry point and
nothing more. Registering the rule no longer requires seeing how it works.

The improvement margin becomes a constant local to the implementation: only
chooseJoinOrder reads it, and the tests that discuss it do so in prose.

Every relocated body is byte-identical to its previous text.
