+++ arangod/Aql/Optimizer/Rule/OptimizeJoinOrder/CMakeLists.txt
@@ -2,5 +2,6 @@ target_sources(arango_aql PRIVATE
   IndexJoinStatistics.cpp
   JoinGraph.cpp
   JoinGraphConstruction.cpp
+  JoinOrderSearch.cpp
   SystemRCostEstimator.cpp
 )
