+++ tests/Aql/JoinGraphTestHelper.h
@@ -27,7 +27,7 @@
 #include "Aql/ExecutionNode/EnumerateCollectionNode.h"
 #include "Aql/ExecutionNode/ExecutionNode.h"
 #include "Aql/ExecutionPlan.h"
-#include "Aql/Optimizer/Rule/OptimizeJoinOrder.h"
+#include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinGraph.h"
 #include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinGraphConstruction.h"
 #include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinCostEstimator.h"
 #include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinStatistics.h"
